feat: add ProtectedRouteElement
