+++ src/components/protected-route-element/protected-route-element.tsx
@@ -1,0 +1,31 @@
+import { useAppSelector } from "@src/hooks/store-hooks";
+import { selectUserStatus } from "@src/store/reducers/profileInfo/selectors";
+import { AuthStatus } from "@src/types/types";
+import { FC, ReactElement, useEffect } from "react";
+import { useNavigate } from "react-router-dom";
+
+interface IProtectedRouteElement {
+  element: ReactElement;
+  protectedPageType: "auth" | "main";
+}
+
+const ProtectedRouteElement: FC<IProtectedRouteElement> = ({
+  element,
+  protectedPageType,
+}) => {
+  const navigate = useNavigate();
+  const userStatus = useAppSelector(selectUserStatus);
+  useEffect(() => {
+    switch (protectedPageType) {
+      case "auth":
+        if (userStatus === AuthStatus.authorized) navigate("/main");
+        break;
+      case "main":
+        if (userStatus === AuthStatus.notAuthorized) navigate("/");
+        break;
+    }
+  }, [navigate, userStatus]);
+  return <div>{element}</div>;
+};
+
+export default ProtectedRouteElement;
