+++ src/hooks/useWindowSize.ts
@@ -1,0 +1,19 @@
+import { useEffect, useState } from "react";
+
+export function useWindowSize(maxWidth: string | number) {
+  const [width, setWidth] = useState(window.innerWidth);
+
+  useEffect(() => {
+    const handleResize = (event: any) => {
+      setWidth(event.target.innerWidth);
+    };
+    window.addEventListener("resize", handleResize);
+    return () => {
+      window.removeEventListener("resize", handleResize);
+    };
+  }, []);
+
+  return width < Number(maxWidth);
+
+  return;
+}
